Fix for recordings lacking a photo reference

+++ mobile/lib/models/recording_info.dart
@@ -3,8 +3,8 @@ import 'package:audio_service/audio_service.dart';
 class RecordingInfo {
   RecordingInfo.fromJson(Map<String, dynamic> jsonData)
       : title = jsonData['acf']['title'] as String,
-        thumbnailPath = (jsonData['acf']['image'] as int).toString(),
-        recordingPath = (jsonData['acf']['file'] as int).toString(),
+        thumbnailPath = jsonData['acf']['image'].toString(),
+        recordingPath = jsonData['acf']['file'].toString(),
         duration = Duration.zero;
 
   final String title;
